# public/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dashboard Management API",
    "version": "1.0.0",
    "description": "A comprehensive API for user authentication and management with role-based permissions",
    "contact": {
      "name": "API Support",
      "email": "app@carik.id"
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message description"
          }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing authentication",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "type": "object",
                  "properties": {
                    "message": {
                      "example": "Invalid or missing authentication token"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "type": "object",
                  "properties": {
                    "message": {
                      "example": "Insufficient permissions to access this resource"
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/v1/ai/chat/completions": {
      "post": {
        "summary": "AI Chat Completions",
        "description": "Proxy endpoint for AI chat completions. Forwards requests to the configured AI service.",
        "tags": [
          "AI"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string",
                    "example": "carik/gpt-6-mini"
                  },
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string",
                          "example": "user"
                        },
                        "content": {
                          "type": "string",
                          "example": "Hello, how are you?"
                        }
                      }
                    }
                  },
                  "temperature": {
                    "type": "number",
                    "example": 0.7
                  },
                  "stream": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response from AI service",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}